add some notes
