+++ src/circuit/README.md
@@ -1,0 +1,16 @@
+- [ ] Basic structure of types in quantum circuit and adiabatic model
+  - [ ] C-NOT
+  - [ ] Hadamard
+  - [ ] Pauli-X, Pauli-Y, Pauli-Z
+  - [ ] TimeOp
+  - [ ] object: `Circuit`
+
+- Quantum circuits and adiabatic computation
+  - [ ] `addgate!`, `removegate!`
+  - [ ] pHamiltonian
+
+- Visualization
+  - [ ] `plot` (circuit)
+
+- Error correction
+  - [ ] CSS code
