+++ src/circuit/gates.jl
@@ -1,7 +1,3 @@
-abstract AbstractModels{N}
-abstract QuCircuit{N}<:AbstractModels{N}
-
-
 ##########################
 # Gates
 ##########################
@@ -70,7 +66,6 @@ end
 CtrlGateUnit{T,N}(gate::Gate{T,N},pos::Vector{Int},ctrl::Vector{Int}) = CtrlGateUnit{T,N}(gate,pos,ctrl)
 
 type HadamardUnit<:AbstractGateUnit{1}
-    gate::AbstractMatrix
     pos::Int
 end
 
@@ -85,42 +80,3 @@ type PhaseUnit<:AbstractGateUnit{1}
 end
 
 bitnum{T,N}(unit::AbstractGateUnit{N}) = N
-
-############################
-# Circuits
-############################
-
-type Circuit{N} <: QuCircuit{N}
-    gates::Array{GateUnit,1}
-end
-
-Circuit(num::Integer,gates::Array{GateUnit,1}) = Circuit{num}(gates)
-Circuit(num::Integer)=Circuit{num}(Array(GateUnit,0))
-
-###########################
-# Stablizer Circuit
-# Circuits with only Hadamard, C-NOT, R (Phase)
-# followed by only one bits measurement
-###########################
-
-type stlzCircuit{N} <: QuCircuit{N}
-    gates::Array{GateUnit,1}
-end
-
-stlzCircuit(num::Integer,gates::Array{GateUnit,1}) = StableCircuit{num}(gates)
-stlzCircuit(num::Integer) = StableCircuit{num}(Array(GateUnit,0))
-
-############################
-# Circuit Constructor
-############################
-
-max(a::Int) = a
-
-function addgate!{T,N,M}(circuit::QuCircuit{N},gate::Gate{T,M},pos::Int...)
-    # Bounds check
-    @assert length(pos)==M+1 "number of qubits do not match"
-    @assert max(pos[2:end]...)<=N
-
-    push!(circuit.gates,GateUnit(gate,ntuple(x->sort(collect(pos))[x],length(pos))...))
-    sort!(circuit.gates,alg=QuickSort,by=x->x.pos[1])
-end
